# Tuning CRC calculations