Modified CRC calculations to reference precalculated results
rather than calculate the CRC for each byte through each call.

Signed-off-by: James Swaro <jswaro@cray.com>

Associated with #461