@jswaro 2016-09-01
This change only affects the RDMA paths. This drops the API time by ~0.06 usec on average. 

Before:
```
# Size        Bandwidth (MB/s)        Latency (us)        Min Lat (us)        Max Lat (us)   Avg Write API(us)      Avg CQ API(us)
8192                   5784.53                1.35                1.35                1.35                0.53                0.80
16384                  7082.86                2.21                2.21                2.21                0.53                1.65
32768                  8036.92                3.89                3.89                3.89                0.53                3.34
65536                  8645.52                7.23                7.23                7.23                0.53                6.68
131072                 9088.14               13.75               13.75               13.75                0.53               13.20
262144                 9281.09               26.94               26.94               26.94                0.53               26.38
524288                 9416.10               53.10               53.10               53.10                0.54               52.54
1048576                9434.76              105.99              105.99              105.99                0.54              105.43
```

After
```
# Size        Bandwidth (MB/s)        Latency (us)        Min Lat (us)        Max Lat (us)   Avg Write API(us)      Avg CQ API(us)
8192                   5783.83                1.35                1.35                1.35                0.47                0.86
16384                  7090.34                2.20                2.20                2.20                0.47                1.71
32768                  8044.57                3.88                3.88                3.88                0.47                3.39
65536                  8615.23                7.25                7.25                7.25                0.48                6.76
131072                 9064.26               13.79               13.79               13.79                0.48               13.29
262144                 9258.01               27.00               27.00               27.00                0.48               26.51
524288                 9392.93               53.23               53.23               53.23                0.48               52.74
1048576                9443.12              105.90              105.90              105.90                0.48              105.40
2097152                9460.47              211.41              211.41              211.41                0.48              210.91
4194304                9230.58              433.34              433.34              433.34                0.48              432.84
```